docs: outline tutorial
