+++ code/webauthn/contracts/deploy/deploy.ts
@@ -1,10 +1,9 @@
 import { utils, Wallet, Provider } from 'zksync-ethers';
-import * as ethers from 'ethers';
 import type { HardhatRuntimeEnvironment } from 'hardhat/types';
 import { Deployer } from '@matterlabs/hardhat-zksync-deploy';
-
 // load env file
 import dotenv from 'dotenv';
+import { ethers } from 'ethers';
 dotenv.config();
 
 const DEPLOYER_PRIVATE_KEY = process.env.WALLET_PRIVATE_KEY || '';
@@ -14,45 +13,52 @@ export default async function (hre: HardhatRuntimeEnvironment) {
   const provider = new Provider(hre.network.config.url);
   const wallet = new Wallet(DEPLOYER_PRIVATE_KEY, provider);
   const deployer = new Deployer(hre, wallet);
-  await deploy(deployer, wallet);
+
+  // deploy the AA Factory & test deploying an account
+  await deployAAFactory(deployer);
+  // deploy the NFT contract
+  await deployMyNFT(deployer);
+  // deploy the Paymaster contract
+  await deployPaymaster(deployer, wallet);
 }
 
-export async function deploy(deployer: Deployer, wallet: Wallet) {
+async function deployAAFactory(deployer: Deployer) {
   const factoryArtifact = await deployer.loadArtifact('AAFactory');
   const aaArtifact = await deployer.loadArtifact('Account');
 
   const factory = await deployer.deploy(factoryArtifact, [utils.hashBytecode(aaArtifact.bytecode)], undefined, [
     aaArtifact.bytecode,
-  ]);
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  ] as any);
   const factoryAddress = await factory.getAddress();
   console.log(`AA factory address: ${factoryAddress}`);
+}
 
-  const aaFactory = new ethers.Contract(factoryAddress, factoryArtifact.abi, wallet);
+async function deployMyNFT(deployer: Deployer) {
+  const artifact = await deployer.loadArtifact('MyNFT');
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const constructorArguments: any[] = [];
+  const contract = await deployer.deploy(artifact, constructorArguments);
+  console.log('NFT CONTRACT ADDRESS: ', await contract.getAddress());
 
-  const owner = Wallet.createRandom();
-  console.log('SC Account owner pk: ', owner.privateKey);
-
-  const salt = ethers.ZeroHash;
-  const tx = await aaFactory.deployAccount(salt, owner.address);
+  const tx = await contract.mintZeek();
   await tx.wait();
+  console.log('DONE MINTING');
+}
 
-  const abiCoder = new ethers.AbiCoder();
-  const accountAddress = utils.create2Address(
-    factoryAddress,
-    await aaFactory.aaBytecodeHash(),
-    salt,
-    abiCoder.encode(['address'], [owner.address])
-  );
+async function deployPaymaster(deployer: Deployer, wallet: Wallet) {
+  const artifact = await deployer.loadArtifact('GeneralPaymaster');
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const constructorArguments: any[] = [];
+  const contract = await deployer.deploy(artifact, constructorArguments);
+  const paymasterAddress = await contract.getAddress();
+  console.log('PAYMASTER CONTRACT ADDRESS: ', paymasterAddress);
 
-  console.log(`SC Account deployed on address ${accountAddress}`);
+  const tx = await wallet.sendTransaction({
+    to: paymasterAddress,
+    value: ethers.parseEther('10'),
+  });
 
-  console.log('Funding smart contract account with some ETH');
-  await (
-    await wallet.sendTransaction({
-      to: accountAddress,
-      value: ethers.parseEther('100'),
-    })
-  ).wait();
-  console.log(`Done!`);
-  return { accountAddress, owner };
+  await tx.wait();
+  console.log('DONE DEPLOYING & FUNDING PAYMASTER');
 }
