--- code/webauthn/contracts/deploy/deployPaymaster.ts
@@ -1,31 +1,0 @@
-import { Wallet, Provider } from 'zksync-ethers';
-import type { HardhatRuntimeEnvironment } from 'hardhat/types';
-import { Deployer } from '@matterlabs/hardhat-zksync-deploy';
-
-// load env file
-import dotenv from 'dotenv';
-import { ethers } from 'ethers';
-dotenv.config();
-
-const DEPLOYER_PRIVATE_KEY = process.env.WALLET_PRIVATE_KEY || '';
-
-export default async function (hre: HardhatRuntimeEnvironment) {
-  // @ts-expect-error target config file which can be testnet or local
-  const provider = new Provider(hre.network.config.url);
-  const wallet = new Wallet(DEPLOYER_PRIVATE_KEY, provider);
-  const deployer = new Deployer(hre, wallet);
-  const artifact = await deployer.loadArtifact('GeneralPaymaster');
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const constructorArguments: any[] = [];
-  const contract = await deployer.deploy(artifact, constructorArguments);
-  const paymasterAddress = await contract.getAddress();
-  console.log('PAYMASTER CONTRACT ADDRESS: ', paymasterAddress);
-
-  const tx = await wallet.sendTransaction({
-    to: paymasterAddress,
-    value: ethers.parseEther('10'),
-  });
-
-  await tx.wait();
-  console.log('DONE DEPLOYING & FUNDING PAYMASTER');
-}
