+++ code/webauthn/frontend/src/components/Layout.tsx
@@ -1,8 +1,6 @@
-import { buttonStyles } from '@/pages';
 import { Inter } from 'next/font/google';
 import Link from 'next/link';
 import React from 'react';
-import { BUTTON_COLORS } from '../../utils/constants';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -12,7 +10,7 @@ export function Layout({ children, isHome }: { children: React.ReactNode; isHome
       {!isHome && (
         <div style={{ margin: '2rem 1rem' }}>
           <Link
-            style={{ ...buttonStyles, background: BUTTON_COLORS[7] }}
+            style={buttonStyles}
             href="/"
           >
             Home
@@ -28,3 +26,17 @@ const styles = {
   padding: '2rem',
   minHeight: '100vh',
 };
+
+const buttonStyles = {
+  background: 'linear-gradient(90deg, #cfecd0, #a0cea7, #9ec0db)',
+  padding: '1rem 2rem',
+  borderRadius: '5px',
+  textDecoration: 'none',
+  cursor: 'pointer',
+  color: 'black',
+  border: '1px solid white',
+  fontSize: '1.2rem',
+  width: '300px',
+  margin: 'auto',
+  textAlign: 'center',
+} as React.CSSProperties;
