--- code/webauthn/frontend/utils/constants.ts
@@ -1,16 +1,0 @@
-export const AA_FACTORY_ADDRESS = '0x<YOUR_AA_FACTORY_ADDRESS>';
-
-export const PAYMASTER_ADDRESS = '0x<YOUR_PAYMASTER_ADDRESS>';
-
-export const NFT_CONTRACT_ADDRESS = '0x<YOUR_NFT_CONTRACT_ADDRESS>';
-
-export const BUTTON_COLORS = [
-  'linear-gradient(90deg, #69b7eb, #b3dbd3, #f4d6db)',
-  'linear-gradient(90deg, #cfecd0, #ffc5ca)',
-  'linear-gradient(90deg, #f598a8, #f6edb2)',
-  'linear-gradient(90deg, #ee5c87, #f1a4b5, #d587b3)',
-  'linear-gradient(90deg, #b9deed, #efefef)',
-  'linear-gradient(90deg, #aea4e3, #d3ffe8)',
-  'linear-gradient(90deg, #faf0cd, #fab397)',
-  'linear-gradient(90deg, #cfecd0, #a0cea7, #9ec0db)',
-];
