+++ code/webauthn/frontend/utils/tx.ts
@@ -2,7 +2,7 @@ import { DEFAULT_GAS_PER_PUBDATA_LIMIT, getPaymasterParams } from 'zksync-ethers
 import { EIP712Signer, type Wallet, type Provider, utils, type types } from 'zksync-ethers';
 import { ethers } from 'ethers';
 import * as accountAbiJSON from '../../contracts/artifacts-zk/contracts/Account.sol/Account.json';
-import { PAYMASTER_ADDRESS } from './constants';
+import { PAYMASTER_ADDRESS } from '@/pages/index';
 
 export async function getTransaction(to: string, from: string, value: string, data: string, provider: Provider) {
   const gasPrice = await provider.getGasPrice();
