+++ content/tutorials/guide-viem/_dir.yml
@@ -1,5 +1,5 @@
 title: Using viem with ZKsync
-featured: true
+featured: false
 authors:
   - name: MatterLabs
     url: https://matter-labs.io
