--- content/tutorials/sign-with-webauthn/10.index.md
@@ -1,24 +1,0 @@
----
-title: Using P256Verify in a Smart Account
-description: Use the P256Verify precompile
----
-
-Smart contract account using P256:
-
-`AAFactory.sol`:
-
-```solidity
-:code-import{filePath="webauthn/contracts/contracts/AAFactory.sol"}
-```
-
-change the end of the `_validateTransaction` function:
-
-```solidity
-:code-import{filePath="webauthn/contracts/contracts/Account.sol:_validateTransaction"}
-```
-
-Here is the entire `Account.sol` file:
-
-```solidity
-:code-import{filePath="webauthn/contracts/contracts/Account.sol"}
-```
