--- content/tutorials/sign-with-webauthn/20.webauthn.md
@@ -1,13 +1,0 @@
----
-title: Using Webauthn
-description: Connect webauthn to your smart account.
-
----
-
-Use webauthn:
-
-Make a new file called `sign.ts`:
-
-```ts
-:code-import{filePath="webauthn/frontend/utils/sign.ts"}
-```
