+++ content/tutorials/signing-transactions-with-webauthn/10.index.md
@@ -1,0 +1,58 @@
+---
+title: Signing Transactions With WebAuthn
+description: Learn how to sign transactions with WebAuthn
+---
+
+## Prerequisites
+
+- Make sure your machine satisfies the [system
+  requirements](https://github.com/matter-labs/era-compiler-solidity/tree/main#system-requirements).
+- You should have [Node.js](https://nodejs.org/en/download) installed.
+- You should have [`era_test_node`](https://docs.zksync.io/build/test-and-debug/in-memory-node#install-and-set-up-era_test_node) installed.
+- If you aren't familiar with paymasters and smart accounts on ZKsync Era, check out the
+[General Paymaster](https://docs.zksync.io/build/start-coding/zksync-101/paymaster) and
+[Native AA Multisig](https://code.zksync.io/tutorials/native-aa-multisig) tutorials.
+
+## What You Will Build
+
+In this tutorial, we will build an app that allows users to sign transactions with WebAuthn.
+
+To break this down, the app will allow users to:
+
+- Deploy a new instance of a smart account.
+- Create a new WebAuthn passkey.
+- Register the WebAuthn passkey as an authorized user of the smart account.
+- Use WebAuthn to Transfer ETH.
+- Use WebAuthn to mint an NFT.
+- Have gasless transactions.
+
+<video width="100%" height="100%" controls>
+  <source src="/videos/webauthn-demo.mp4" type="video/mp4">
+  Your browser does not support the video tag.
+</video>
+
+You can find the completed code inside the [Community Code repo](https://github.com/zkSync-Community-Hub/community-code/code/webauthn).
+
+## Introduction to WebAuthn
+
+WebAuthn, short for Web Authentication API, is an API built-in to your browser that allows your app to register and authenticate users.
+It allows users to authenticate using biometric data (such as Apple's TouchID) or hardware devices (such as a Yubikey).
+
+### Ceremonies
+
+There are two steps, or "ceremonies", for using WebAuthn: registration and authentication.
+
+1. The first step, registration, is when the user creates a new key pair specific to the application they are using.
+2. The second step, authentication, is used to send a challenge for WebAuthn to sign, and receive a signature in return.
+
+### Signatures
+
+WebAuthn uses a different signature scheme than what is typically used in blockchain.
+
+WebAuthn uses the secp256r1 elliptic curve for signing, while standard accounts in blockchain use the secp256k1 curve.
+
+Without a helper function to handle verifying secp256r1 signatures, it's very difficult to integrate a smart account with WebAuthn.
+
+Luckily, we can utilize the P256Verify precompile to verify that a given WebAuthn signature matches a known public key.
+
+In the next section, we will create the contracts for the demo app.
