+++ content/tutorials/signing-transactions-with-webauthn/20.contracts.md
@@ -1,0 +1,213 @@
+---
+title: Building the Contracts
+description: Build the contracts for the WebAuthn app.
+
+---
+
+## Project Setup
+
+Make a new folder for the project called `zksync-webauthn` and navigate to that folder in your terminal:
+
+```shell
+mkdir zksync-webauthn
+cd zksync-webauthn
+```
+
+We can start by creating the contracts for the smart account, paymaster, and NFT.
+
+```shell
+zksync-cli create contracts --template hardhat_solidity --project contracts
+```
+
+Enter the private key for a pre-configured rich wallet:
+
+```shell
+? Private key of the wallet responsible for deploying contracts (optional)
+0x7726827caac94a7f9e1b160f7ea819f172f7b6f9d2a97f992c38edeab82d4110
+```
+
+Once that is done, move into the `contracts` folders and delete the template contracts, scripts, and tests:
+
+```shell
+cd contracts
+rm -rf ./contracts/*
+rm -rf ./deploy/*
+rm -rf ./test/*
+```
+
+## Contracts
+
+Open the `zksync-webauthn` folder in your preferred IDE.
+
+We will be creating 4 contracts:
+
+- `AAFactory.sol`: a factory contract to deploy new instances of smart accounts.
+- `Account.sol`: the contract for the smart account.
+- `GeneralPaymaster.sol`: a general paymaster contract to sponsor transactions.
+- `MyNFT.sol`: a generic NFT contract.
+
+Create the new contract files in the `contracts/contracts` folder:
+
+```shell
+touch {contracts/AAFactory.sol,contracts/Account.sol,contracts/GeneralPaymaster.sol,contracts/MyNFT.sol}
+```
+
+Copy and paste the code below into respective files:
+
+### AA Factory
+
+::drop-panel
+::panel{label="AAFactory.sol"}
+
+  ```solidity [contracts/AAFactory.sol]
+  :code-import{filePath="webauthn/contracts/contracts/AAFactory.sol"}
+  ```
+
+::
+::
+
+### Smart Contract Account
+
+::drop-panel
+::panel{label="Account.sol"}
+
+```solidity [contracts/Account.sol]
+:code-import{filePath="webauthn/contracts/contracts/Account.sol"}
+```
+
+::
+::
+
+### Paymaster Contract
+
+::drop-panel
+::panel{label="GeneralPaymaster.sol"}
+
+```solidity [contracts/GeneralPaymaster.sol]
+:code-import{filePath="webauthn/contracts/contracts/GeneralPaymaster.sol"}
+```
+
+::
+::
+
+### NFT Contract
+
+::drop-panel
+::panel{label="MyNFT.sol"}
+
+```solidity [contracts/MyNFT.sol]
+:code-import{filePath="webauthn/contracts/contracts/MyNFT.sol"}
+```
+
+::
+::
+
+## Understanding the WebAuthn Signature Validation
+
+The smart contract account implements a way to validate signatures created with WebAuthn.
+
+The `_validateTransaction` function first checks to see if there is a valid secp256k1 signature from the account owner,
+just like the example in the TUTORIAL TODO: LINK.
+
+If the transaction is not signed by the standard account owner, the next check is to see if the transaction contains a valid WebAuthn signature.
+
+```solidity [contracts/Account.sol]
+:code-import{filePath="webauthn/contracts/contracts/Account.sol:_validateTransaction"}
+```
+
+The contract contains a state variable `r1Owner` to store the the public key of the WebAuthn account that the owner can register.
+
+```solidity [contracts/Account.sol]
+:code-import{filePath="webauthn/contracts/contracts/Account.sol:r1Owner"}
+```
+
+```solidity [contracts/Account.sol]
+:code-import{filePath="webauthn/contracts/contracts/Account.sol:validateWebAuthnSignature"}
+```
+
+### Validating the Signature
+
+To validate the signature, the `_validateWebAuthnSignature` first extracts the `authenticatorData`, `clientData`, and `rs` values from the decoded signature.
+
+It extracts the WebAuthn challenge from the `clientData`, and checks to see if it matches the transaction hash.
+This means that the exact transaction data that is submitted to the contract matches the transaction data signed by the WebAuthn account.
+
+It checks for the malleability of the signature,
+and checks to see if the WebAuthn user flags are set.
+
+Finally, it calls the `callVerifier` function to call the `P256Verify` precompile.
+
+```solidity [contracts/Account.sol]
+:code-import{filePath="webauthn/contracts/contracts/Account.sol:_validateWebAuthnSignature"}
+```
+
+The `callVerifier` function encodes the input for into the correct format and calls the `P256Verify` precompile.
+
+If the returned output is empty, the validation evaluated to false.
+This means that the public key derived from the WebAuthn signature does not match the registered `r1Owner` public key.
+
+The output data should be 1 (in 32 bytes format) if the signature verification process succeeds, or nothing if it fails.
+
+```solidity [contracts/Account.sol]
+:code-import{filePath="webauthn/contracts/contracts/Account.sol:callVerifier"}
+```
+
+## Deploying
+
+### Using a Local Node
+
+Open a new terminal and start a local in-memory node with `era_test_node`:
+
+```shell
+era_test_node run
+```
+
+Next, modify the `defaultNetwork` in the `hardhat.config.ts` file to deploy to the local in-memory node, and change the version of Solidity to `0.8.20`:
+
+```ts
+  defaultNetwork: "inMemoryNode",
+```
+
+```ts
+  solidity: {
+    version: "0.8.20",
+  },
+```
+
+### Adding the Deploy Script
+
+Create a new file inside the `deploy` folder called `deploy.ts`:
+
+```shell
+touch deploy/deploy.ts
+```
+
+Copy and paste the code below. This script will:
+
+- deploy the `AAFactory` contract
+- deploy the NFT contract and mint an NFT
+- deploy the paymaster contract and send funds to it from the preconfigured rich account.
+
+::drop-panel
+::panel{label="deploy.ts"}
+
+```solidity [deploy/deploy.ts]
+:code-import{filePath="webauthn/contracts/deploy/deploy.ts"}
+```
+
+::
+::
+
+### Running the Deploy Script
+
+Finally, compile and deploy the contracts with:
+
+```shell
+npm run compile
+npm run deploy
+```
+
+Save the output of this command, as we will use these deployed contract addresses in the frontend.
+
+That's all for the contracts!
+In the next section, we will build the frontend.
