+++ content/tutorials/signing-transactions-with-webauthn/30.frontend.md
@@ -1,0 +1,227 @@
+---
+title: Building the Frontend
+description: Build the frontend for the WebAuthn app.
+
+---
+
+## Frontend Setup
+
+We will use Next.js to build the frontend.
+
+Move out of your `contract` folder and into the `zksync-webauthn` folder.
+Make a new Next.js project for the frontend using the configuration below:
+
+```shell
+npx create-next-app@latest
+```
+
+```txt
+✔ What is your project named? … frontend
+✔ Would you like to use TypeScript? Yes
+✔ Would you like to use ESLint? Yes
+✔ Would you like to use Tailwind CSS? No
+✔ Would you like to use `src/` directory? Yes
+✔ Would you like to use App Router? (recommended) No
+✔ Would you like to customize the default import alias (@/*)? No
+```
+
+Move into your projects and install the dependencies below:
+
+```shell
+cd frontend
+npm install @simplewebauthn/browser@10.0.0 @simplewebauthn/server@10.0.1 cbor@9.0.2 ethers@5.7.2 zksync-ethers@5.1.0
+```
+
+## Home Page
+
+Replace your `frontend/src/pages/index.tsx` file with the code below:
+
+::drop-panel
+::panel{label="index.tsx"}
+
+```ts [frontend/src/pages/index.tsx]
+:code-import{filePath="webauthn/frontend/src/pages/index.tsx"}
+```
+
+::
+::
+
+## Layout
+
+Create a `src/components` folder for the `Layout` component.
+
+::drop-panel
+::panel{label="layout.tsx"}
+
+```ts [frontend/src/components/Layout.tsx]
+:code-import{filePath="webauthn/frontend/src/components/Layout.tsx"}
+```
+
+::
+::
+
+## Hooks
+
+Next, create a `src/hooks` folder and add the hooks below to manage the created smart account.
+
+The `useAccount` hook saves and retrieves the account address of the user's smart account using local storage.
+
+::drop-panel
+::panel{label="useAccount.tsx"}
+
+```ts [frontend/src/hooks/useAccount.tsx]
+:code-import{filePath="webauthn/frontend/src/hooks/useAccount.tsx"}
+```
+
+::
+::
+
+The `useWallet` hook temporarily saves the full wallet owner of the deployed smart contract account to use for registering a passkey.
+
+::drop-panel
+::panel{label="useWallet.tsx"}
+
+```ts [frontend/src/hooks/useWallet.tsx]
+:code-import{filePath="webauthn/frontend/src/hooks/useWallet.tsx"}
+```
+
+::
+::
+
+## API
+
+The `get-registration-options` route is used to generate the options input object for the WebAuthn registration ceremony.
+
+These options tell the browser how to configure the registration,
+including setting the name of the passkey, requiring multi-factor authentication, and more.
+You can learn more about the options available for registration in the [`simplewebauthn` docs](https://simplewebauthn.dev/docs/packages/server#1-generate-registration-options).
+
+::drop-panel
+::panel{label="get-registration-options.ts"}
+
+```ts [frontend/src/pages/api/get-registration-options.ts]
+:code-import{filePath="webauthn/frontend/src/pages/api/get-registration-options.ts"}
+```
+
+::
+::
+
+The `generate-authentication-options` route is used to generate the options input object for the WebAuthn authentication ceremony.
+You can learn more about the options available for authentication in the [`simplewebauthn` docs](https://simplewebauthn.dev/docs/packages/server#1-generate-authentication-options).
+
+The challenge passed in through the request body represents the transaction data that would normally be signed by a wallet like Metamask.
+In this case, however, instead of sending to a wallet we are sending the transaction to the browser to sign with WebAuthn.
+
+::drop-panel
+::panel{label="generate-authentication-options.ts"}
+
+```ts [frontend/src/pages/api/generate-authentication-options.ts]
+:code-import{filePath="webauthn/frontend/src/pages/api/generate-authentication-options.ts"}
+```
+
+::
+::
+
+## Utils
+
+Make a new folder called `frontend/utils` where you can add some utility functions for your app.
+
+Create a new file inside the `utils` folder called `string.ts` to help handle some string conversions you will need.
+
+::drop-panel
+::panel{label="string.ts"}
+
+```ts [frontend/utils/string.ts]
+:code-import{filePath="webauthn/frontend/utils/string.ts"}
+```
+
+::
+::
+
+Next, create a file called `tx.ts`.
+This file contains helper functions to get data for a transaction
+and register a new passkey in a smart account.
+
+::drop-panel
+::panel{label="tx.ts"}
+
+```ts [frontend/utils/tx.ts]
+:code-import{filePath="webauthn/frontend/utils/tx.ts"}
+```
+
+::
+::
+
+The last file in the `utils` folder is called `webauthn.ts`,
+which contains helper functions to start the WebAuthn authentication process,
+sign and send transactions.
+
+::drop-panel
+::panel{label="webauthn.ts"}
+
+```ts [frontend/utils/tx.ts]
+:code-import{filePath="webauthn/frontend/utils/webauthn.ts"}
+```
+
+::
+::
+
+## Pages
+
+Inside the `frontend/src/pages` folder create four new pages: `create-account.tsx`, `register.tsx`, `mint.tsx`, and `transfer.tsx`:
+
+::drop-panel
+::panel{label="create-account.tsx"}
+
+```ts [frontend/src/pages/create-account.tsx]
+:code-import{filePath="webauthn/frontend/src/pages/create-account.tsx"}
+```
+
+::
+::
+
+::drop-panel
+::panel{label="register.tsx"}
+
+```ts [frontend/src/pages/register.tsx]
+:code-import{filePath="webauthn/frontend/src/pages/register.tsx"}
+```
+
+::
+::
+
+::drop-panel
+::panel{label="mint.tsx"}
+
+```ts [frontend/src/pages/mint.tsx]
+:code-import{filePath="webauthn/frontend/src/pages/mint.tsx"}
+```
+
+::
+::
+
+::drop-panel
+::panel{label="transfer.tsx"}
+
+```ts [frontend/src/pages/transfer.tsx]
+:code-import{filePath="webauthn/frontend/src/pages/transfer.tsx"}
+```
+
+::
+::
+
+## Running the App
+
+Start the frontend app by running the command below in the `frontend` folder:
+
+```shell
+npm run dev
+```
+
+You can now try out the app!
+
+Follow the order of the buttons on the home page:
+
+1. Start by creating a new account.
+1. Next, register a new passkey.
+1. Finally, you can use the registered passkey to transfer ETH and mint an NFT!
